Share CUDA client routing helpers
