+++ codegen/codegen.py
@@ -1680,6 +1680,7 @@ def main():
             "#include <unordered_map>\n"
             "#include <vector>\n\n"
             '#include "gen_api.h"\n\n'
+            '#include "client_routing.h"\n'
             '#include "rpc.h"\n\n'
             "extern int rpc_size();\n"
             "extern conn_t *rpc_client_get_connection(unsigned int index);\n"
@@ -1687,30 +1688,9 @@ def main():
             'extern "C" void lupine_deep_cache_reset(const void *key);\n'
             'extern "C" void *lupine_deep_cache_add(const void *key, '
             "size_t bytes);\n\n"
-            "struct lupine_route {\n"
-            "    int kind;\n"
-            "    conn_t *conn;\n"
-            "};\n\n"
             'extern "C" CUresult lupine_cuInit_multi(unsigned int flags);\n'
             'extern "C" CUresult lupine_cuDeviceGetCount_multi(int *count);\n'
             'extern "C" CUresult lupine_cuDeviceGet_multi(CUdevice *device, int ordinal);\n'
-            'extern "C" lupine_route lupine_route_for_default();\n'
-            'extern "C" lupine_route lupine_route_for_device(CUdevice *device);\n'
-            'extern "C" lupine_route lupine_route_for_current_context();\n'
-            'extern "C" lupine_route lupine_route_for_context(CUcontext ctx);\n'
-            'extern "C" lupine_route lupine_route_for_module(CUmodule module);\n'
-            'extern "C" lupine_route lupine_route_for_library(CUlibrary library);\n'
-            'extern "C" lupine_route lupine_route_for_function(CUfunction function);\n'
-            'extern "C" lupine_route lupine_route_for_stream(CUstream stream);\n'
-            'extern "C" lupine_route lupine_route_for_event(CUevent event);\n'
-            'extern "C" lupine_route lupine_route_for_memory_pool(CUmemoryPool pool);\n'
-            'extern "C" lupine_route lupine_route_for_graph(CUgraph graph);\n'
-            'extern "C" lupine_route lupine_route_for_graph_node(CUgraphNode node);\n'
-            'extern "C" lupine_route lupine_route_for_graph_exec(CUgraphExec exec);\n'
-            'extern "C" lupine_route lupine_route_for_deviceptr(CUdeviceptr ptr);\n'
-            'extern "C" bool lupine_route_is_local(lupine_route route);\n'
-            'extern "C" conn_t *lupine_route_remote_conn(lupine_route route);\n'
-            'extern "C" void *lupine_real_cuda_symbol(const char *name);\n'
             'extern "C" conn_t *lupine_rpc_conn_for_device(CUdevice *device);\n'
             'extern "C" conn_t *lupine_rpc_conn_for_current_context();\n'
             'extern "C" conn_t *lupine_rpc_conn_for_context(CUcontext ctx);\n'
@@ -1732,18 +1712,6 @@ def main():
             'extern "C" void lupine_note_graph_exec_owner(CUgraphExec exec, conn_t *conn);\n'
             'extern "C" void lupine_note_deviceptr_owner(CUdeviceptr ptr, conn_t *conn);\n\n'
             'extern "C" void lupine_note_deviceptr_allocation(CUdeviceptr ptr, size_t size, conn_t *conn);\n\n'
-            'extern "C" void lupine_note_context_owner_route(CUcontext ctx, lupine_route route);\n'
-            'extern "C" void lupine_note_module_owner_route(CUmodule module, lupine_route route);\n'
-            'extern "C" void lupine_note_library_owner_route(CUlibrary library, lupine_route route);\n'
-            'extern "C" void lupine_note_function_owner_route(CUfunction function, lupine_route route);\n'
-            'extern "C" void lupine_note_stream_owner_route(CUstream stream, lupine_route route);\n'
-            'extern "C" void lupine_note_event_owner_route(CUevent event, lupine_route route);\n'
-            'extern "C" void lupine_note_memory_pool_owner_route(CUmemoryPool pool, lupine_route route);\n'
-            'extern "C" void lupine_note_graph_owner_route(CUgraph graph, lupine_route route);\n'
-            'extern "C" void lupine_note_graph_node_owner_route(CUgraphNode node, lupine_route route);\n'
-            'extern "C" void lupine_note_graph_exec_owner_route(CUgraphExec exec, lupine_route route);\n'
-            'extern "C" void lupine_note_deviceptr_owner_route(CUdeviceptr ptr, lupine_route route);\n\n'
-            'extern "C" void lupine_note_deviceptr_allocation_route(CUdeviceptr ptr, size_t size, lupine_route route);\n\n'
             'extern "C" void lupine_forget_deviceptr_owner(CUdeviceptr ptr);\n\n'
             'extern "C" void lupine_record_library_kernel(CUkernel kernel, CUlibrary library, const char *name, lupine_route route);\n\n'
             'extern "C" void lupine_record_module_function(CUfunction function, CUmodule module, const char *name, lupine_route route);\n\n'
@@ -1772,22 +1740,6 @@ def main():
             'extern "C" CUresult lupine_cuDeviceCanAccessPeer_multi(int *canAccessPeer, CUdevice dev, CUdevice peerDev);\n'
             'extern "C" CUresult lupine_cuCtxEnablePeerAccess_multi(CUcontext peerContext, unsigned int flags);\n'
             'extern "C" CUresult lupine_cuCtxDisablePeerAccess_multi(CUcontext peerContext);\n\n'
-            "template <typename Fn> static Fn lupine_real_cuda_fn(const char *name) {\n"
-            "    return reinterpret_cast<Fn>(lupine_real_cuda_symbol(name));\n"
-            "}\n\n"
-            "template <typename Fn, typename... Args>\n"
-            "static bool lupine_call_local_cuda_if_routed(lupine_route route,\n"
-            "                                             const char *symbol,\n"
-            "                                             CUresult *result,\n"
-            "                                             Args... args) {\n"
-            "    if (!lupine_route_is_local(route)) {\n"
-            "        return false;\n"
-            "    }\n"
-            "    auto real = lupine_real_cuda_fn<Fn>(symbol);\n"
-            "    *result = real == nullptr ? CUDA_ERROR_DEVICE_UNAVAILABLE\n"
-            "                              : real(args...);\n"
-            "    return true;\n"
-            "}\n\n"
         )
         for function, annotation, operations, metadata in functions_with_annotations:
             # We don't generate client function definitions for client-disabled
