+++ codegen/gen_client.cpp
@@ -14,6 +14,7 @@
 
 #include "gen_api.h"
 
+#include "client_routing.h"
 #include "rpc.h"
 
 extern int rpc_size();
@@ -22,31 +23,9 @@ extern void rpc_close(conn_t *conn);
 extern "C" void lupine_deep_cache_reset(const void *key);
 extern "C" void *lupine_deep_cache_add(const void *key, size_t bytes);
 
-struct lupine_route {
-  int kind;
-  conn_t *conn;
-};
-
 extern "C" CUresult lupine_cuInit_multi(unsigned int flags);
 extern "C" CUresult lupine_cuDeviceGetCount_multi(int *count);
 extern "C" CUresult lupine_cuDeviceGet_multi(CUdevice *device, int ordinal);
-extern "C" lupine_route lupine_route_for_default();
-extern "C" lupine_route lupine_route_for_device(CUdevice *device);
-extern "C" lupine_route lupine_route_for_current_context();
-extern "C" lupine_route lupine_route_for_context(CUcontext ctx);
-extern "C" lupine_route lupine_route_for_module(CUmodule module);
-extern "C" lupine_route lupine_route_for_library(CUlibrary library);
-extern "C" lupine_route lupine_route_for_function(CUfunction function);
-extern "C" lupine_route lupine_route_for_stream(CUstream stream);
-extern "C" lupine_route lupine_route_for_event(CUevent event);
-extern "C" lupine_route lupine_route_for_memory_pool(CUmemoryPool pool);
-extern "C" lupine_route lupine_route_for_graph(CUgraph graph);
-extern "C" lupine_route lupine_route_for_graph_node(CUgraphNode node);
-extern "C" lupine_route lupine_route_for_graph_exec(CUgraphExec exec);
-extern "C" lupine_route lupine_route_for_deviceptr(CUdeviceptr ptr);
-extern "C" bool lupine_route_is_local(lupine_route route);
-extern "C" conn_t *lupine_route_remote_conn(lupine_route route);
-extern "C" void *lupine_real_cuda_symbol(const char *name);
 extern "C" conn_t *lupine_rpc_conn_for_device(CUdevice *device);
 extern "C" conn_t *lupine_rpc_conn_for_current_context();
 extern "C" conn_t *lupine_rpc_conn_for_context(CUcontext ctx);
@@ -71,33 +50,6 @@ extern "C" void lupine_note_deviceptr_owner(CUdeviceptr ptr, conn_t *conn);
 
 extern "C" void lupine_note_deviceptr_allocation(CUdeviceptr ptr, size_t size,
                                                  conn_t *conn);
-
-extern "C" void lupine_note_context_owner_route(CUcontext ctx,
-                                                lupine_route route);
-extern "C" void lupine_note_module_owner_route(CUmodule module,
-                                               lupine_route route);
-extern "C" void lupine_note_library_owner_route(CUlibrary library,
-                                                lupine_route route);
-extern "C" void lupine_note_function_owner_route(CUfunction function,
-                                                 lupine_route route);
-extern "C" void lupine_note_stream_owner_route(CUstream stream,
-                                               lupine_route route);
-extern "C" void lupine_note_event_owner_route(CUevent event,
-                                              lupine_route route);
-extern "C" void lupine_note_memory_pool_owner_route(CUmemoryPool pool,
-                                                    lupine_route route);
-extern "C" void lupine_note_graph_owner_route(CUgraph graph,
-                                              lupine_route route);
-extern "C" void lupine_note_graph_node_owner_route(CUgraphNode node,
-                                                   lupine_route route);
-extern "C" void lupine_note_graph_exec_owner_route(CUgraphExec exec,
-                                                   lupine_route route);
-extern "C" void lupine_note_deviceptr_owner_route(CUdeviceptr ptr,
-                                                  lupine_route route);
-
-extern "C" void lupine_note_deviceptr_allocation_route(CUdeviceptr ptr,
-                                                       size_t size,
-                                                       lupine_route route);
 
 extern "C" void lupine_forget_deviceptr_owner(CUdeviceptr ptr);
 
@@ -147,22 +99,6 @@ extern "C" CUresult lupine_cuDeviceCanAccessPeer_multi(int *canAccessPeer,
 extern "C" CUresult lupine_cuCtxEnablePeerAccess_multi(CUcontext peerContext,
                                                        unsigned int flags);
 extern "C" CUresult lupine_cuCtxDisablePeerAccess_multi(CUcontext peerContext);
-
-template <typename Fn> static Fn lupine_real_cuda_fn(const char *name) {
-  return reinterpret_cast<Fn>(lupine_real_cuda_symbol(name));
-}
-
-template <typename Fn, typename... Args>
-static bool lupine_call_local_cuda_if_routed(lupine_route route,
-                                             const char *symbol,
-                                             CUresult *result, Args... args) {
-  if (!lupine_route_is_local(route)) {
-    return false;
-  }
-  auto real = lupine_real_cuda_fn<Fn>(symbol);
-  *result = real == nullptr ? CUDA_ERROR_DEVICE_UNAVAILABLE : real(args...);
-  return true;
-}
 
 CUresult cuInit(unsigned int Flags) { return lupine_cuInit_multi(Flags); }
